Extract block from block_queue
Extract a class block from block_queue service
